Buffer switching finally works. Audio also works

+++ adsr.cpp
@@ -4,33 +4,36 @@ class adsr
 {
 public:
     bool key;
-    double attack;
-    double decay;
-    double sustain;
-    double sustainLevel;
-    double release;
+    float attack;
+    float decay;
+    float sustain;
+    float sustainLevel;
+    float release;
     int sample;
     int sampleRate;
 
-    double proces(){
-        if (sample <= sampleRate*attack){
-            double a = 1/(sample);
+    float proces()
+    {
+        if (sample <= sampleRate * attack)
+        {
+            float a = 1 / (sample);
             return (sample * a);
         }
-        if (sample < sampleRate*(attack + decay)){
+        if (sample < sampleRate * (attack + decay))
+        {
             int b = 1;
-            double a = (sustainLevel - b)/(sample);
+            float a = (sustainLevel - b) / (sample);
             return (sample * a + b);
         }
-        if (sample <= sampleRate*(attack + decay + sustain)){
-            return sustainLevel; 
+        if (sample <= sampleRate * (attack + decay + sustain))
+        {
+            return sustainLevel;
         }
-        if (sample < sampleRate*(attack + decay + sustain + release)){
-            double b = sustainLevel;
-            double a = (-b)/(sample);
+        if (sample < sampleRate * (attack + decay + sustain + release))
+        {
+            float b = sustainLevel;
+            float a = (-b) / (sample);
             return (sample * a + b);
         }
     }
 };
-
-
