+++ headers/filters/baseEffect.hpp
@@ -3,24 +3,21 @@
 class BaseEffect
 {
 private:
-
     bool active = false;
     bool debug = false;
 
 public:
-
     BaseEffect();
     BaseEffect(bool active);
     BaseEffect(bool active, bool debug);
 
     virtual ~BaseEffect();
 
-    virtual void calcParams (double td);
-    virtual double process (double sample); 
+    virtual void calcParams(float td);
+    virtual float process(float sample);
 
-    bool isActive ();
+    bool isActive();
 
     void activate();
     void deactivate();
 };
-
