+++ src/notes.cpp
@@ -2,18 +2,18 @@
 #include <math.h>
 #include <string>
 
-NotesCalculator::NotesCalculator (int numOfSemitonsofBase, double baseFreq) {
+NotesCalculator::NotesCalculator(int numOfSemitonsofBase, float baseFreq)
+{
     this->semitonesFromStart = numOfSemitonsofBase;
     this->baseFreq = baseFreq;
-    
 }
 
-double NotesCalculator::GetFreqOfNote (int n) 
+float NotesCalculator::GetFreqOfNote(int n)
 {
     return this->baseFreq * std::pow(1.05946309, n);
 }
 
-int NotesCalculator::GetSemitonesFromNoteName (std::string noteName)
+int NotesCalculator::GetSemitonesFromNoteName(std::string noteName)
 {
     return 1;
 }
